+++ lib/src/material/material_progress_bar.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:video_player/video_player.dart';
 
 import '../../src/chewie_progress_colors.dart';
@@ -28,9 +27,9 @@ class MaterialVideoProgressBar extends StatelessWidget {
   Widget build(BuildContext context) {
     return VideoProgressBar(
       controller,
-      barHeight: 2,
+      barHeight: 10,
       handleHeight: 6,
-      drawShadow: false,
+      drawShadow: true,
       colors: colors,
       onDragEnd: onDragEnd,
       onDragStart: onDragStart,